pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Agent with tool calling, LangGraph, and open-source LLMs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
   "langgraph",
   "transformers",
   "accelerate",
   "sentence-transformers",
   "chroma-hnswlib",
   "httpx",
   "deepeval",
   "pytest",
   "pyyaml",
   "python-dotenv",
   "fastapi>=0.120.0",
   "uvicorn>=0.38.0",
   "langchain>=1.0.2",
   "wikipedia>=1.4.0",
   "langchain-community>=0.4.1",
   "langchain-huggingface>=1.0.0",
]

